rename: change userInfo class package
