--- src/main/java/com/example/backend/dto/user/UserInfo.java
@@ -1,5 +1,0 @@
-package com.example.backend.dto.user;
-
-public abstract class UserInfo {
-
-}
