+++ src/main/java/com/example/backend/service/user/userInfo/UserInfo.java
@@ -1,0 +1,5 @@
+package com.example.backend.service.user.userInfo;
+
+public abstract class UserInfo {
+
+}
